[#15] feat: auth error code 추가 및 일부 적용
